FIX: Better test and reorg
